# docs/swagger.yaml
basePath: /v1
definitions:
...
  candidate.AddPersonalInfoRequest:
    properties:
      address:
        type: string
      email:
        type: string
      name:
        type: string
      phone:
        type: string
    required:
    - email
    - name
    type: object
  candidate.AddProjectRequest:
    properties:
      category:
        type: string
      description:
        type: string
      project_link:
        type: string
      project_name:
        type: string
    required:
    - project_link
    - project_name
    type: object
  candidate.AddSkillRequest:
    properties:
      skill:
        type: string
    required:
    - skill
    type: object
  candidate.UpdatePersonalInfoRequest:
    properties:
      address:
        type: string
      email:
        type: string
      name:
        type: string
      phone:
        type: string
    required:
    - email
    - name
    type: object
...
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
      status:
        type: string
    type: object
host: dz-jobs-api-production.up.railway.app
info:
  contact: {}
  description: This is the API documentation for the DZ Jobs portal.
  title: DZ Jobs API
  version: "1.0"
paths:
...
  /candidates/{candidate_id}/personal-info:
    delete:
      description: Delete personal information for a candidate
      parameters:
      - description: Candidate ID
        in: path
        name: candidate_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: Delete personal information
      tags:
      - Candidates - Personal Info
    get:
      description: Get personal information for a candidate by candidate ID
      parameters:
      - description: Candidate ID
        in: path
        name: candidate_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: Get personal information by candidate ID
      tags:
      - Candidates - Personal Info
    post:
      consumes:
      - application/json
      description: Add personal information for a candidate
      parameters:
      - description: Candidate ID
        in: path
        name: candidate_id
        required: true
        type: string
      - description: Personal Info request
        in: body
        name: personal_info
        required: true
        schema:
          $ref: '#/definitions/candidate.AddPersonalInfoRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: Add personal information
      tags:
      - Candidates - Personal Info
    put:
      consumes:
      - application/json
      description: Update personal information for a candidate
      parameters:
      - description: Candidate ID
        in: path
        name: candidate_id
        required: true
        type: string
      - description: Personal Info request
        in: body
        name: personal_info
        required: true
        schema:
          $ref: '#/definitions/candidate.UpdatePersonalInfoRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: Update personal information
      tags:
      - Candidates - Personal Info
  /candidates/{candidate_id}/portfolio:
    get:
      description: Get all projects for a candidate by candidate ID
      parameters:
      - description: Candidate ID
        in: path
        name: candidate_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: Get projects by candidate ID
      tags:
      - Candidates - Portfolio
    post:
      consumes:
      - application/json
      description: Add a new project for a candidate
      parameters:
      - description: Candidate ID
        in: path
        name: candidate_id
        required: true
        type: string
      - description: Project request
        in: body
        name: project
        required: true
        schema:
          $ref: '#/definitions/candidate.AddProjectRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: Add a new project
      tags:
      - Candidates - Portfolio
  /candidates/{candidate_id}/portfolio/{project_id}:
    delete:
      description: Delete a project by candidate ID and project ID
      parameters:
      - description: Candidate ID
        in: path
        name: candidate_id
        required: true
        type: string
      - description: Project ID
        in: path
        name: project_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: Delete project
      tags:
      - Candidates - Portfolio
  /candidates/{candidate_id}/skills:
    get:
      description: Get all skills for a candidate by candidate ID
      parameters:
      - description: Candidate ID
        in: path
        name: candidate_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: Get skills by candidate ID
      tags:
      - Candidates - Skills
    post:
      consumes:
      - application/json
      description: Add a new skill for a candidate
      parameters:
      - description: Candidate ID
        in: path
        name: candidate_id
        required: true
        type: string
      - description: Skill request
        in: body
        name: Skill
        required: true
        schema:
          $ref: '#/definitions/candidate.AddSkillRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: Add a new skill
      tags:
      - Candidates - Skills
  /candidates/{candidate_id}/skills/{skill_name}:
    delete:
      description: Delete a skill by candidate ID and skill name
      parameters:
      - description: Candidate ID
        in: path
        name: candidate_id
        required: true
        type: string
      - description: Skill name
        in: path
        name: skill_name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: Delete skill
      tags:
      - Candidates - Skills
...
swagger: "2.0"